Assembler calculation service

+++ HeighwayDragon/Calculations/ServiceLayer/IDragonCalculator.cs
@@ -8,9 +8,9 @@ namespace HeighwayDragon.Calculations.ServiceLayer
 	public interface IDragonCalculator
 	{
 		bool ChooseTransformation();
-		double CalculateX1(double x, double y);
-		double CalculateX2(double x, double y);
-		double CalculateY1(double x, double y);
-		double CalculateY2(double x, double y);
+		float CalculateX1(float x, float y);
+		float CalculateX2(float x, float y);
+		float CalculateY1(float x, float y);
+		float CalculateY2(float x, float y);
 	}
 };
